Consolidate query types
